Renamed files * RankScene to RankingScene * RankWorker to RankingWorker * RankInteractor to RankingInteractor * RankModel to RankingModel * RankPresenter to RankingPresenter * RankView to RankingView
