--- project/app/src/main/java/com/nexte/nexte/RankScene/RankView.kt
@@ -1,9 +1,0 @@
-package com.nexte.nexte.RankScene
-
-/**
- * Created by albino on 02/04/18.
- */
-
-interface RankDisplayLogic {
-    fun displayRankInScreen(viewModel: RankModel.ViewModel)
-}
