+++ project/app/src/main/java/com/nexte/nexte/RankingScene/RankingInteractor.kt
@@ -1,19 +1,19 @@
-package com.nexte.nexte.RankScene
+package com.nexte.nexte.RankingScene
 
 /**
  * Created by albino on 02/04/18.
  */
 
 interface RankBusinessLogic {
-    fun getPlayersRanksForScene(request: RankModel.Request)
+    fun getPlayersRanksForScene(request: RankingModel.Request)
 }
 
 class RankInteractor : RankBusinessLogic {
 
-    var worker = RankWorker()
+    var worker = RankingWorker()
     var presenter: RankPresentationLogic? = null
 
-    override fun getPlayersRanksForScene(request: RankModel.Request) {
+    override fun getPlayersRanksForScene(request: RankingModel.Request) {
         worker.getUsersInRank(request){ response ->
             presenter?.presentRank(response)
         }
