+++ project/app/src/main/java/com/nexte/nexte/RankingScene/RankingView.kt
@@ -1,0 +1,9 @@
+package com.nexte.nexte.RankingScene
+
+/**
+ * Created by albino on 02/04/18.
+ */
+
+interface RankDisplayLogic {
+    fun displayRankInScreen(viewModel: RankingModel.ViewModel)
+}
